refactor: remove collection-mounted tests and integrate functionality into collection tests
- Deleted the collection-mounted test file and moved relevant tests into the collection test file.
- Updated the readCollectionEntryRow function to handle both key and payload arguments.
- Added new tests for useItem invalidation, collection invalidation, repeated invalidations, hook remount behavior, and collection hydration.
- Enhanced test coverage for async storage efficiency in indexedDB.
